requirements.txt
altair==4.2.0
anyio==3.6.2
attrs==22.1.0
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
decorator==5.1.1
entrypoints==0.4
fastapi==0.86.0
gitdb==4.0.9
GitPython==3.1.29
h11==0.14.0
idna==3.4
imbalanced-learn==0.9.1
imblearn==0.0
importlib-metadata==5.0.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.0
MarkupSafe==2.1.1
numpy==1.23.4
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
protobuf==3.20.3
pyarrow==10.0.0
pydantic==1.10.2
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
rich==12.6.0
scikit-learn==1.1.3
scipy==1.9.3
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.20.4
streamlit==1.14.0
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
uvicorn==0.19.0
validators==0.20.0
watchdog==2.1.9
zipp==3.10.0
